storage_volume(s): move Read to read.go; fix type_id + expand filters
Address follow-on review feedback on the storage volume data sources.

- Move the Read() method and its helpers out of datasource.go into a dedicated
  read.go for both hpe_morpheus_storage_volume and hpe_morpheus_storage_volumes,
  leaving datasource.go with the schema/metadata scaffolding (per maintainer
  direction to adopt the per-operation file split for these data sources).

- storage_volumes: map type_id from the top-level typeId rather than the nested
  type object (sv.Type.Id). This matches the singular data source and the filter
  logic and keeps type_id populated when the API omits the nested type object
  but sets the top-level field.

- storage_volumes: expand the filter switch (and the matching OneOf validator)
  to cover the remaining useful string fields - type_name, volume_type,
  volume_name, volume_path, category, disk_type, disk_mode, device_name,
  storage_group_id, storage_group_name, source, external_id, internal_id,
  unique_id, namespace, storage_profile - so filtering by them returns matches
  instead of silently yielding an empty result.
