+++ morpheus/framework/datasources/storagevolumes/schema_gen.go
@@ -246,10 +246,10 @@ func StorageVolumesDataSourceSchema(ctx context.Context) schema.Schema {
 					Attributes: map[string]schema.Attribute{
 						"name": schema.StringAttribute{
 							Required:            true,
-							Description:         "The field to filter on. Valid names are: name, type_id, type_code, datastore_id, datastore_name, cloud_id, cloud_name, storage_server_id, storage_server_name, status, uuid, provision_type, ref_type, ref_id, pool_name.",
-							MarkdownDescription: "The field to filter on. Valid names are: name, type_id, type_code, datastore_id, datastore_name, cloud_id, cloud_name, storage_server_id, storage_server_name, status, uuid, provision_type, ref_type, ref_id, pool_name.",
+							Description:         "The field to filter on. Valid names are: name, type_id, type_code, type_name, volume_type, volume_name, volume_path, category, disk_type, disk_mode, device_name, datastore_id, datastore_name, cloud_id, cloud_name, storage_server_id, storage_server_name, storage_group_id, storage_group_name, status, uuid, provision_type, ref_type, ref_id, pool_name, source, external_id, internal_id, unique_id, namespace, storage_profile.",
+							MarkdownDescription: "The field to filter on. Valid names are: name, type_id, type_code, type_name, volume_type, volume_name, volume_path, category, disk_type, disk_mode, device_name, datastore_id, datastore_name, cloud_id, cloud_name, storage_server_id, storage_server_name, storage_group_id, storage_group_name, status, uuid, provision_type, ref_type, ref_id, pool_name, source, external_id, internal_id, unique_id, namespace, storage_profile.",
 							Validators: []validator.String{
-								stringvalidator.OneOf("name", "type_id", "type_code", "datastore_id", "datastore_name", "cloud_id", "cloud_name", "storage_server_id", "storage_server_name", "status", "uuid", "provision_type", "ref_type", "ref_id", "pool_name"),
+								stringvalidator.OneOf("name", "type_id", "type_code", "type_name", "volume_type", "volume_name", "volume_path", "category", "disk_type", "disk_mode", "device_name", "datastore_id", "datastore_name", "cloud_id", "cloud_name", "storage_server_id", "storage_server_name", "storage_group_id", "storage_group_name", "status", "uuid", "provision_type", "ref_type", "ref_id", "pool_name", "source", "external_id", "internal_id", "unique_id", "namespace", "storage_profile"),
 							},
 						},
 						"values": schema.SetAttribute{
